.history/requirements_20231108023025.txt
django >=4, <4.1
djangorestFramework
pyYAML
requests
django-cors-headers